adding method to add data with csv

+++ src/main/java/com/sanger/muni/dto/liquidacion/NewConceptoDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class NewConceptoDto {
 
     private Long id;
-    private Short cantidad;
+    private Float cantidad;
     private Double importe;
 
 }
